aoc_2022_09/aoc_2022_09_test.go
package aoc_2022_09

import (
	"strings"
	"testing"
)

const Example1 = `R 4
U 4
L 3
D 1
R 4
D 1
L 5
R 2`

const Example2 = `R 5
U 8
L 8
D 3
R 17
D 10
L 25
U 20`

func TestPart1Examples(t *testing.T) {
	egs := map[string]Part1Result{
		Example1: 13,
	}

	for in, want := range egs {
		got := DoPart1(ParseInput(strings.NewReader(in)))
		if got != want {
			t.Errorf("DoPart1(%q) got %v, wanted %v", in, got, want)
		}
	}
}

func TestPart2Examples(t *testing.T) {
	egs := map[string]Part2Result{
		Example2: 36,
	}

	for in, want := range egs {
		got := DoPart2(ParseInput(strings.NewReader(in)))
		if got != want {
			t.Errorf("DoPart2(%q) got %v, wanted %v", in, got, want)
		}
	}
}

func TestPart1(t *testing.T) {
	const want = Part1Want
	got := Part1()
	if got != want {
		t.Errorf("Part1() got %v, wanted %v", got, want)
	}
}

func TestPart2(t *testing.T) {
	const want = Part2Want
	got := Part2()
	if got != want {
		t.Errorf("Part2() got %v, wanted %v", got, want)
	}
}
